pannelli bf senza retro ok

+++ scaricaPannello.php
@@ -8,44 +8,55 @@
 
     if($configurazione == "BF" || $configurazione == "bf")
     {
-        $query3="EXEC insert_incollaggio_lana";
-        $result3=sqlsrv_query($conn,$query3);
-        if($result3==FALSE)
-            die("error2".$query3);
+        $query30="SELECT * FROM dw_incollaggio.dbo.pannelli_linea WHERE id_distinta=$id_distinta AND faccia='retro'";
+        $result30=sqlsrv_query($conn,$query30);
+        if($result30==FALSE)
+            die("error2".$query30);
         else
         {
-            $query4="INSERT INTO [dbo].[pannelli_prodotti] ([id_distinta],[faccia],[bancale],[tempo_di_percorrenza])
-                    SELECT $id_distinta,'retro',(SELECT id_bancale FROM dbo.anagrafica_bancali WHERE (stato = 'aperto')),(SELECT DATEDIFF(second, dbo.pannelli_linea.dataOra, GETDATE()) AS DateDiff FROM dbo.pannelli_linea INNER JOIN dbo.anagrafica_stazioni ON dbo.pannelli_linea.stazione = dbo.anagrafica_stazioni.id_stazione WHERE (dbo.pannelli_linea.id_distinta = $id_distinta) AND (dbo.pannelli_linea.faccia = 'retro') AND (dbo.anagrafica_stazioni.nome = 'caricamento'))";
-            $result4=sqlsrv_query($conn,$query4);
-            if($result4==TRUE)
+            $rows = sqlsrv_has_rows( $result30 );
+            if ($rows === true)
             {
-                $query7="INSERT INTO [dbo].[distinta_pannelli_prodotti]
-                                ([pannello_prodotto]
-                                ,[stazione]
-                                ,[utente]
-                                ,[dataOra]
-                                ,[faccia])
-                        SELECT (SELECT MAX(id_pannello_prodotto) FROM pannelli_prodotti WHERE id_distinta = $id_distinta AND faccia='retro'),[stazione],[utente],[dataOra],[faccia] 
-                        FROM pannelli_linea 
-                        WHERE id_distinta=$id_distinta AND faccia='retro'";
-                $result7=sqlsrv_query($conn,$query7);
-                if($result7==FALSE)
-                    die("error2".$query7);
+                $query3="EXEC insert_incollaggio_lana";
+                $result3=sqlsrv_query($conn,$query3);
+                if($result3==FALSE)
+                    die("error2".$query3);
                 else
                 {
-                    $query5="DELETE FROM pannelli_linea WHERE id_distinta=$id_distinta AND faccia='retro'";
-                    $result5=sqlsrv_query($conn,$query5);
-                    if($result5==FALSE)
-                        die("error2".$query5);
+                    $query4="INSERT INTO [dbo].[pannelli_prodotti] ([id_distinta],[faccia],[bancale],[tempo_di_percorrenza])
+                            SELECT $id_distinta,'retro',(SELECT MAX(id_bancale) AS id_bancale FROM dbo.anagrafica_bancali WHERE (stato = 'aperto')),(SELECT DATEDIFF(second, dbo.pannelli_linea.dataOra, GETDATE()) AS DateDiff FROM dbo.pannelli_linea INNER JOIN dbo.anagrafica_stazioni ON dbo.pannelli_linea.stazione = dbo.anagrafica_stazioni.id_stazione WHERE (dbo.pannelli_linea.id_distinta = $id_distinta) AND (dbo.pannelli_linea.faccia = 'retro') AND (dbo.anagrafica_stazioni.nome = 'caricamento'))";
+                    $result4=sqlsrv_query($conn,$query4);
+                    if($result4==TRUE)
+                    {
+                        $query7="INSERT INTO [dbo].[distinta_pannelli_prodotti]
+                                        ([pannello_prodotto]
+                                        ,[stazione]
+                                        ,[utente]
+                                        ,[dataOra]
+                                        ,[faccia])
+                                SELECT (SELECT MAX(id_pannello_prodotto) FROM pannelli_prodotti WHERE id_distinta = $id_distinta AND faccia='retro'),[stazione],[utente],[dataOra],[faccia] 
+                                FROM pannelli_linea 
+                                WHERE id_distinta=$id_distinta AND faccia='retro'";
+                        $result7=sqlsrv_query($conn,$query7);
+                        if($result7==FALSE)
+                            die("error2".$query7);
+                        else
+                        {
+                            $query5="DELETE FROM pannelli_linea WHERE id_distinta=$id_distinta AND faccia='retro'";
+                            $result5=sqlsrv_query($conn,$query5);
+                            if($result5==FALSE)
+                                die("error2".$query5);
+                        }
+                    }
+                    else
+                        die("error".$query4);
                 }
             }
-            else
-                die("error".$query4);
         }
     }
 
     $query2="INSERT INTO [dbo].[pannelli_prodotti] ([id_distinta],[faccia],[bancale],[tempo_di_percorrenza])
-            SELECT $id_distinta,'$faccia',(SELECT id_bancale FROM dbo.anagrafica_bancali WHERE (stato = 'aperto')),(SELECT DATEDIFF(second, dbo.pannelli_linea.dataOra, GETDATE()) AS DateDiff FROM dbo.pannelli_linea INNER JOIN dbo.anagrafica_stazioni ON dbo.pannelli_linea.stazione = dbo.anagrafica_stazioni.id_stazione WHERE (dbo.pannelli_linea.id_distinta = $id_distinta) AND (dbo.pannelli_linea.faccia = '$faccia') AND (dbo.anagrafica_stazioni.nome = 'caricamento'))";
+            SELECT $id_distinta,'$faccia',(SELECT MAX(id_bancale) AS id_bancale FROM dbo.anagrafica_bancali WHERE (stato = 'aperto')),(SELECT DATEDIFF(second, dbo.pannelli_linea.dataOra, GETDATE()) AS DateDiff FROM dbo.pannelli_linea INNER JOIN dbo.anagrafica_stazioni ON dbo.pannelli_linea.stazione = dbo.anagrafica_stazioni.id_stazione WHERE (dbo.pannelli_linea.id_distinta = $id_distinta) AND (dbo.pannelli_linea.faccia = '$faccia') AND (dbo.anagrafica_stazioni.nome = 'caricamento'))";
     $result2=sqlsrv_query($conn,$query2);
     if($result2==TRUE)
     {
